backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.5.0
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
python-dotenv>=1.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-generativeai>=0.3.0
gunicorn
requests
yfinance>=0.2.0
certifi>=2023.0.0
openai>=1.0.0
pydub>=0.25.0
scipy>=1.11.0
scikit-learn>=1.3.0
deepgram-sdk>=3.0.0
google-cloud-texttospeech>=2.14.0
boto3>=1.28.0
sentence-transformers>=2.2.0
torch>=2.0.0
